fix(ios): change GenerationResponse.remainingCredits to Double?
The generate-story edge function returns remainingCredits as a fractional
number (e.g., 53.75 after a $0.25 charge from a fractional credit balance).
iOS struct declared Int?; JSONDecoder failed with 'the data couldn't be read
because it isn't in the correct format.'

creditCostCharged is already Double? for the same reason (PR #263 migrated
estimatedCredits Int->Double). remainingCredits was missed in that migration.

Fixes 'Could not parse server response' on Generate action when the user has
a fractional remaining credit balance.

+++ CathedralOSApp/Services/GenerationRequestDTO.swift
@@ -206,7 +206,7 @@ struct GenerationResponse: Codable {
     /// The number of credits charged for this generation (present on success).
     let creditCostCharged: Double?
     /// The number of credits remaining after this generation (present on success).
-    let remainingCredits: Int?
+    let remainingCredits: Double?
 
     // MARK: Rate limiting
     /// Seconds the client should wait before retrying after a rate_limited error.
@@ -239,7 +239,7 @@ struct GenerationResponse: Codable {
         requiredCredits     = try c.decodeIfPresent(Int.self, forKey: .requiredCredits)
         availableCredits    = try c.decodeIfPresent(Int.self, forKey: .availableCredits)
         creditCostCharged   = try c.decodeIfPresent(Double.self, forKey: .creditCostCharged)
-        remainingCredits    = try c.decodeIfPresent(Int.self, forKey: .remainingCredits)
+        remainingCredits    = try c.decodeIfPresent(Double.self, forKey: .remainingCredits)
         retryAfterSeconds   = try c.decodeIfPresent(Int.self, forKey: .retryAfterSeconds)
         status              = try c.decode(String.self, forKey: .status)
         errorMessage        = try c.decodeIfPresent(String.self, forKey: .errorMessage)
